(#75) 【音乐持久化】 音乐专辑列表、单曲评论数据库实现

+++ app/src/main/java/com/zhiyicx/thinksnsplus/data/source/local/MusicAlbumListBeanGreenDaoImpl.java
@@ -44,7 +44,7 @@ public class MusicAlbumListBeanGreenDaoImpl extends CommonCacheImpl<MusicAlbumLi
 
     @Override
     public MusicAlbumListBean getSingleDataFromCache(Long primaryKey) {
-        return mMusicAlbumListBeanDao.load(primaryKey.intValue());
+        return mMusicAlbumListBeanDao.load(primaryKey);
     }
 
     @Override
